be_ai/requirements.txt
pymongo
python-dotenv
fastapi
uvicorn[standard]
requests
pydantic
pytesseract
opencv-python
Pillow
numpy
PyMuPDF
google-generativeai
chromadb
sentence-transformers
